fix: split auth config for Edge middleware compatibility
Prisma uses Node.js modules (node:path, node:buffer) which aren't
supported in Edge Runtime. Split auth into:
- auth.config.ts: Edge-compatible config (callbacks, pages, session)
- auth.ts: Full config with Prisma Credentials provider (Node.js only)

Middleware now imports only the lightweight auth.config.ts.
