+++ lib/auth.ts
@@ -3,6 +3,7 @@ import Credentials from "next-auth/providers/credentials";
 import { PrismaClient } from "@/lib/generated/prisma/client";
 import { PrismaPg } from "@prisma/adapter-pg";
 import { compare } from "bcryptjs";
+import { authConfig } from "./auth.config";
 
 function getPrisma() {
   const connectionString = process.env.DATABASE_URL;
@@ -12,48 +13,7 @@ function getPrisma() {
 }
 
 export const { handlers, signIn, signOut, auth } = NextAuth({
-  pages: {
-    signIn: "/admin/login",
-  },
-  session: {
-    strategy: "jwt",
-  },
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user) {
-        token.id = user.id;
-        token.role = (user as { role?: string }).role;
-      }
-      return token;
-    },
-    async session({ session, token }) {
-      if (session.user) {
-        session.user.id = token.id as string;
-        (session.user as { role?: string }).role = token.role as string;
-      }
-      return session;
-    },
-    async authorized({ auth: session, request }) {
-      const { pathname } = request.nextUrl;
-
-      // Public routes - no auth needed
-      if (
-        pathname.startsWith("/q/") ||
-        pathname.startsWith("/api/growth-os/tracking") ||
-        pathname.startsWith("/api/growth-os/webhook") ||
-        pathname.startsWith("/api/growth-os/quotes/track")
-      ) {
-        return true;
-      }
-
-      // Admin routes need authentication
-      if (pathname.startsWith("/admin") && !pathname.startsWith("/admin/login")) {
-        return !!session?.user;
-      }
-
-      return true;
-    },
-  },
+  ...authConfig,
   providers: [
     Credentials({
       name: "credentials",
